Refactore eventbus
